requirements.txt
Django==5.0.6
djangorestframework==3.15.2
PyJWT==2.8.0
